requirements.txt
requests
python-dotenv
mysql-connector-python
pandas
matplotlib
scikit-learn
numpy
ttkbootstrap
